requirements.txt
python-nmap
requests
beautifulsoup4
colorama